move fullscreen stuff into a FullScreenHandler add Full Screen option under preferences|display correct code analysis error in preferences.xml
