+++ app/src/main/java/com/ds/avare/MainActivity.java
@@ -29,6 +29,7 @@ import android.widget.TabHost.TabSpec;
 import android.widget.TextView;
 
 import com.ds.avare.storage.Preferences;
+import com.ds.avare.utils.FullScreenHandler;
 import com.ds.avare.utils.Helper;
  
 /**
@@ -44,6 +45,7 @@ public class MainActivity extends TabActivity {
     HorizontalScrollView mScrollView;
     int      mScrollWidth;
     Preferences mPref;
+    FullScreenHandler fsHandler;
 
     // Tab panels that can display at the bottom of the screen. These manifest as 
     // separate display panes with their own intent to handle the content. Each one
@@ -138,7 +140,8 @@ public class MainActivity extends TabActivity {
         	setupTab(new TextView(this), getString(R.string.Tools), new Intent(this, SatelliteActivity.class), getIntent());
         }
 
-        Helper.addFullScreenVisibilityChangeListener(mTabHost);
+        fsHandler = new FullScreenHandler(mPref.isFullScreen());
+        fsHandler.addFullScreenVisibilityChangeListener(mTabHost);
     }
     
     /**
@@ -204,7 +207,7 @@ public class MainActivity extends TabActivity {
         super.onWindowFocusChanged(hasFocus);
 
         if(hasFocus) {
-            Helper.activateFullScreenFlags(getWindow().getDecorView());
+            fsHandler.activateFullScreenFlags(getWindow().getDecorView());
         }
     }
