+++ app/src/main/res/xml/preferences.xml
@@ -241,6 +241,11 @@ authors: zkhan, jlmcgraw
             android:key="@string/NightMode"
             android:summary="@string/NightModeSummary"
             android:title="@string/NightModeLabel" />
+        <CheckBoxPreference
+            android:defaultValue="false"
+            android:key="@string/FullScreen"
+            android:summary="@string/FullScreenSummary"
+            android:title="@string/FullScreenLabel" />
         <com.ds.avare.utils.ListPreferenceWithSummary
             android:defaultValue="100000"
             android:entries="@array/ADSBTrafficFilterAltitude"
